feat: add temporal slider
