+++ package.json
@@ -20,6 +20,7 @@
     "react-scripts": "5.0.1",
     "styled-components": "^5.3.5",
     "styled-reset": "^4.4.1",
+    "swiper": "^8.2.4",
     "typescript": "^4.4.2",
     "web-vitals": "^2.1.0"
   },
